`dask/optimize.py -> dask/optimization.py`
- Renames `dask/optimize.py` to `dask/optimization.py`
- Adds deprecated functions in `dask/optimize.py` to forward to
`dask/optimization.py`.
- Add attributes to `dask.base.optimize` function to forward to
deprecated methods, which allows this function to be imported in the
top-level namespace.
