+++ dask/array/optimization.py
@@ -7,7 +7,7 @@ import numpy as np
 from .core import getter, getter_nofancy, getter_inline
 from ..compatibility import zip_longest
 from ..core import flatten, reverse_dict
-from ..optimize import cull, fuse, inline_functions
+from ..optimization import cull, fuse, inline_functions
 from ..utils import ensure_dict
