+++ dask/multiprocessing.py
@@ -7,7 +7,7 @@ import sys
 
 from .local import get_async  # TODO: get better get
 from .context import _globals
-from .optimize import fuse, cull
+from .optimization import fuse, cull
 
 import cloudpickle
